Various bug fixes & quality changes Fixed issue with ui where player can press left or right arrow to exit out of 'OK' prompt without ending player turn.
Also fixed a small bug with 'Flame of Wala' spell which caused it to skip the message and 'OK' prompt for it,
which was caused by an overlooked line in the player.cpp file which relinquished the player's turn upon burning the enemy,
rather than upon pressing enter key on 'OK' prompt.
Additionally, decided to make the enemy not receive burn damage on the same turn that it was inflicted burn.

Finally, allowed for player name to have spaces by using the getline method to get the whole line,
instead of cin which ignored anything after the first space.