Implemented guard + damage rounding & misc fixes Turned the wait option into a guard option which blocks half the damage received from the enemy. This included creating cases for when each enemy attacks to incorporate the damage reduction from the guard, and damage messages informing the player of the reduced damage.
As part of this I also reworked the way damage is applied,
so that decimal values produced as a result of damage and block multipliers are rounded down.
This also included fixing a previous bug where the damage message had values rounded down, while the actual damage was rounded up.

Additionally, there were some fixes made revolving around quitting the game and dying.
Previously when dying or when quitting using the 'q' key, control would be returned to the main menu,
but the main menu wouldn't print the titlescreen text again, resulting in a blank screen.
When dying, the 'you died' text immediately got shifted up by blank lines making room for the titlescreen which never prints.
The blank titlescreen was fixed by printing the titlescreen again immediately when control is returned to the main menu.
Also, the 'you died' text has a pause afterwards now for better pacing before printing the blank lines and the titlescreen.
In addition, the titlescreen can now be returned to with the escape key as well in addition to the q key.