Coded remaining enemies of second part + balancing Implemented the four remaining enemies of the second part. This mostly consisted of their attack functions as well as adding some lines in main to implement an effect that allows attacking twice. Once the enemies were completed, I fixed up the code that creates the objects of those enemies, as well as fixing a bug in the first wave of enemies which accessed index -1 of an array when i was 0, possibly responsible for a segfault.
Additionally, I wrote the conditions to allow one additional item to appear if the player is in the second part of the game
by increasing the range of the randomly generated number passed to the switch statement for deciding on item drops.

Finally, I buffed various items, increasing the heal of yams, and spirit restoration of brush berries,
as well as making items that increase max health and spirit caps to also restore them by that amount when used.
Also, some moves and spells were tweaked as well.