Added Ailment Shield + various fixes + balancing Introduced and implemented the ailment shield mechanic which prevents the player from receiving a status ailment, getting used up in the process. The player gets an ailment shield from using cleanse or eating a wild orange while they don't have an ailment. In the enemy attack methods, if the enemy's attack causes an ailment, the ailment shield prevents the player from receiving an ailment and then is turned off. It also prevents ailments the player may receive from their own moves, getting used up in the damage calculations in their attacking method. Throughout the duration that the player has the ailment shield, print outs of the menu have "*Shielded*" written in the ailment position.
Various bug fixes were made as well. One bug was caused by a typo in the name of the eagle when setting it as the enemy, causing it to not be set.
Another bug fix was to do with the OK screen that pops up when the player doesn't have enough spirit for a spell, requiring two enter presses to close.
This was fixed by avoiding the OK screen altogether to be consistent with other input which don't end the turn, and instead displaying the message there.
Also changed it so an OK screen shows up when the player uses an item instead of a pause.

Also added additional text on the end of the attacks of dingos to make it more clear to the player that they attack twice.
After this, some lines of code were added to allow enemies that attack twice to also receive burn damage twice.

Also fixed a bug with the attack accuracy of some of the new enemies
where a greater than sign instead of a less than sign was used by accident in the accuracy check, reversing the accuracy.

Additionally, there were several buffs to the player such as increasing the Heal spell from 25 to 30 health.
The cleanse spell now also heals 20 health, and the wild orange item heals 30 health.
The spirit cost from elder heal was reduced from 30 to 25, and the damge of nargun's fist was increased by 10.
Also, wrath of wambeen does 5 more damage and burns, and now has unique text when used.

The gap printed out between the player's health and spirit was decreased in order to more closely match the spacing between the player name and health.