Added final enemy + bug fixes + balancing Filled in the methods for the last unfinished enemy. This consisted mainly of writing the code for its attack method, including deciding on which moves it would use, some of which was similar to other enemy attack methods and could be reused such as generating a random number to choose the attack, and another one for accuracy. To make the enemy a little more systematic than others, there were some extra booleans to prevent overuse of a certain move, as well as setting specific conditions for a heal to occur. The main code was then updated to include the enemy which included creating it, as well as breaking out of the while loop when it is beaten, and giving a unique message upon its defeat.
In addition to this, there were several bug fixes, starting with that of moves that incapacitate the player being able to be chained into one another.
This was solved by using a boolean and choosing a different move if the moves were used twice in a row.

Another bug was that revolving around enemies that move twice taking burn damage a second time.
This mechanic wasn't originally implemented properly and caused the text of the enemies second attack to be overwritten,
as well as other issues such as the enemy not dying when their health was reduced to 0 from the second burn and taking another player turn to actually die.
This was addresses by having the burn damage and associated code happen before the enemies second attack as to not overwrite it,
and used a boolean to make sure the enemy wouldn't move again before the death check.

There was also a bug letting the player move twice if they triggered the dialogue about not having enough spirit for a spell.
This was caused by a previous change attempting to remove the 'OK' prompt from this exchange and go straight back to the battle screen.
However, there previously were lines setting it back to the player's turn which should have been removed which were there because the 'OK' screen usually ends the player's turn.
In addition to this, the 'insufficientSpirit' boolean added for this in the previous fix was never set back to false after it was done,
causing all actions to go to the insufficient spirit conditional which doesn't end the turn, inadvertently granting the player an extra turn.

Additionally, some code was written to address the issue where the 'enter' key is pressed during the pause on the poison damage screens, burn damage screens, and screens where the enemy gets an extra attack,
and gets buffered, causing the player to unintentionally enter the attack menu and/or choose moves.
This was done by removing the pause and instead adding blocking character reads wherever these screens popped up, so that the screen is removed by the player instead of by waiting, preventing the buffering of 'enter',
and then setting character reads back to being non-blocking during the player's turn.

A smaller fix was the addition of some carriage returns & newlines in a few of the enemy attack messages to stop them from running on too long on one line.

Finally, there were some balancing changes done to the player.
These were mostly buffs, but there was one nerf. They are as follows:

- Rainbow Restoration now also grants the player an ailment shield when used without a status ailment.

- Hefty Club no longer uses up the player's ailment shield to prevent sluggishness when the enemy dies on the same turn.

- The Brush Berry now fully restores the player's spirit, instead of just by 50.

- The Bush Tomato now increases max health by 15, rather than 10.

- Elder heal now only costs 20 spirit to cast, rather than 25.

- Wambeen's Wrath now costs 10 more spirit to cast.